requirements.txt
python-dotenv==1.0.0
azure-identity==1.14.0
azure-mgmt-dns==8.0.0
requests==2.31.0
six==1.17.0

